change structure and clean
