+++ app/build.gradle
@@ -75,7 +75,6 @@ dependencies {
 
     //chart
     implementation 'com.github.madrapps:plot:0.1.1'
-
     implementation 'com.github.PhilJay:MPAndroidChart:3.1.0'
 
     // retrofit
@@ -94,7 +93,7 @@ dependencies {
     // Google APIs
     implementation 'com.google.android.gms:play-services-location:20.0.0'
 
-    // roomdb
+    // room db
     def roomVersion = "2.5.0-alpha02"
     implementation "androidx.room:room-runtime:$roomVersion"
     annotationProcessor "androidx.room:room-compiler:$roomVersion"
