+++ app/src/main/java/com/example/airquality/services/notification/Notification.kt
@@ -1,6 +1,5 @@
 package com.example.airquality.services.notification
 
-import android.app.AlarmManager
 import android.app.NotificationChannel
 import android.app.NotificationManager
 import android.app.PendingIntent
@@ -9,7 +8,6 @@ import android.content.Intent
 import androidx.core.app.NotificationCompat
 import com.example.airquality.MainActivity
 import com.example.airquality.R
-import java.util.*
 
 class Notification(var context: Context, var title: String, var message: String) {
     private val channelId: String = "HelloNotification"
@@ -26,13 +24,12 @@ class Notification(var context: Context, var title: String, var message: String)
 
         val intent = Intent(context, MainActivity::class.java)
 
-        val pendingIntent = PendingIntent.getActivity(context, 0, intent, PendingIntent.FLAG_MUTABLE)
+        PendingIntent.getActivity(context, 0, intent, PendingIntent.FLAG_MUTABLE)
 
         notificationBuilder = NotificationCompat.Builder(context, channelId)
 
         notificationBuilder
             .setSmallIcon(R.drawable.alarm)
-            //.addAction(R.drawable.ic_launcher_background, "Open Message", pendingIntent)
             .setContentTitle(title)
             .setContentText(message)
             .setAutoCancel(true)
